requirements.txt
asgiref==3.11.1
beautifulsoup4==4.12.3
Brotli==1.0.9
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
cssselect2==0.7.0
dj-database-url==3.1.2
Django==4.0.2
django-cleanup==6.0.0
django-crum==0.7.9
django-environ==0.9.0
django-filter==22.1
django-user-agents==0.4.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
fonttools==4.38.0
gunicorn==25.1.0
html5lib==1.1
idna==3.7
importlib-metadata==5.1.0
Markdown==3.4.1
packaging==26.0
Pillow==9.1.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.5.0
pyphen==0.13.1
pytz==2022.6
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.2
tinycss2==1.2.1
typing_extensions==4.15.0
tzdata==2022.1
ua-parser==0.10.0
urllib3==2.2.1
user-agents==2.2.0
weasyprint==57.1
webencodings==0.5.1
whitenoise==6.12.0
zipp==3.11.0
zopfli==0.2.2
